update controller logic, move bussiness process to model, simplify algorithm
