+++ app/Http/Controllers/LoginController.php
@@ -1,8 +1,7 @@
 <?php
 
-namespace App\Http\Controllers\Login;
+namespace App\Http\Controllers;
 
-use App\Http\Controllers\Controller;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\RedirectResponse;
