+++ app/Http/Controllers/Mahasiswa/KHSController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Mahasiswa;
 
+use App\Http\Controllers\Controller;
 use App\Models\KHS;
 use App\Http\Requests\StoreKHSRequest;
 use App\Http\Requests\UpdateKHSRequest;
