+++ app/Http/Controllers/Mahasiswa/MahasiswaTaskController.php
@@ -1,8 +1,9 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Mahasiswa;
 
 use Illuminate\Http\Request;
+use App\Http\Controllers\Controller;
 use App\Models\Mahasiswa;
 
 class MahasiswaTaskController extends Controller
