+++ app/Http/Controllers/Mahasiswa/PKLController.php
@@ -1,23 +1,23 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Mahasiswa;
 
-use App\Models\Skripsi;
-use App\Http\Requests\StoreSkripsiRequest;
-use App\Http\Requests\UpdateSkripsiRequest;
+use App\Http\Controllers\Controller;
+use App\Models\PKL;
+use App\Http\Requests\StorePKLRequest;
+use App\Http\Requests\UpdatePKLRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Storage;
 
-class SkripsiController extends Controller
+class PKLController extends Controller
 {
     /**
      * Display a listing of the resource.
      */
-    public function index()
-    {   
+    public function index() 
+    {
         $mahasiswa = auth()->user()->mahasiswa;
-       
-        $dataSkripsi = $mahasiswa->skripsi;
+        $dataPKL = $mahasiswa->pkl;
         $dataKHS = $mahasiswa->khs;
         $SKSk = 0;
         $n = 0;
@@ -31,8 +31,8 @@ class SkripsiController extends Controller
             $is_eligible = False;
         }
 
-        return view('mahasiswa.skripsi.index', [
-            'dataSkripsi' => $dataSkripsi,
+        return view('mahasiswa.pkl.index', [
+            'dataPKL' => $dataPKL,
             'is_eligible' => $is_eligible,
         ]);
     }
@@ -46,18 +46,18 @@ class SkripsiController extends Controller
         $validasi = 0;
         // dd($request->status_old);
         if ($status == ""){
-            return redirect('/skripsi');
+            return redirect('/pkl');
         }
         //else
         $rules = [ 
             'semester' => 'required',
-            'tanggal_sidang' => 'required',
+            'tanggal_seminar' => 'required',
             'nilai' => 'required',
         ];
-        if ($request->scan_bass_old == null) {
-            $rules['scan_bass'] = 'required|mimes:pdf|max:10000';
+        if ($request->scan_basp_old == null) {
+            $rules['scan_basp'] = 'required|mimes:pdf|max:10000';
         }else{
-            Storage::delete($request->scan_bass_old);
+            Storage::delete($request->scan_basp_old);
         }
         $validatedData = $request->validate($rules);
 
@@ -66,22 +66,23 @@ class SkripsiController extends Controller
             $validatedData['nama'] = $nama;
             $validatedData['status'] = $status;
             $validatedData['validasi'] = $validasi;
-            if($request->scan_bass != null){
-                $validatedData ["scan_bass"] = $request->file('scan_bass')->store('private/skripsi');
+            if($request->scan_basp != null){
+                $validatedData ["scan_basp"] = $request->file('scan_basp')->store('private/pkl');
             }
-            Skripsi::create($validatedData);
+            PKL::create($validatedData);
         }
         else {
             // $validatedData['nim'] = $nim; 1
             // $validatedData['nama'] = $nama; 2
             $validatedData['status'] = $status;
             // $validatedData['validasi'] = $validasi; 3
-            if($request->scan_bass != null){
-                $validatedData ["scan_bass"] = $request->file('scan_bass')->store('private/skripsi');
+            if($request->scan_basp != null){
+                $validatedData ["scan_basp"] = $request->file('scan_basp')->store('private/pkl');
             }
-            Skripsi::where("nim", $nim)->update($validatedData);
+            PKL::where("nim", $nim)->update($validatedData);
         }
 
-        return redirect('/skripsi')->with('success', "Data Skripsi Berhasil Diubah!");
-     }
+        return redirect('/pkl')->with('success', "Data PKL Berhasil Diubah!");
+    }
 }
+
