+++ app/Http/Controllers/Operator/DepartemenController.php
@@ -1,7 +1,8 @@
 <?php
 
-namespace App\Http\Controllers;
+namespace App\Http\Controllers\Operator;
 
+use App\Http\Controllers\Controller;
 use App\Models\User;
 use App\Models\Departemen;
 use App\Http\Requests\StoreDepartemenRequest;
